feat(controllers/loans): complete loans endpoint
built the endpoint for getting all loans within application. This is a
rather simple  api, but it was very fun to decipher the logic needed. I
had got to built a view(viewSaccoLoans), then called it from another
view(viewLoanApplications) and finnaly called it from a stored procedure
(detailedLoans) to return the data in the format that I needed.
Amazing stuff!