feat(frontend): work on login and signup logic
able to allow login and signup from the frontend side. Everything works
as intended for the most part. However there is a noticable delay when
you first login/signup, unsure where the error is coming fom, but will
make it work later on.

TODO: pressure is creasing as deadline approaches. dont forget to build the database scripts in the backend.