chore(frontend): tried integrating the frontend
alot of the request to the backend do not work as expected. Mainly
because of how the values are structured from the backend, hence
destructuring them raises alot of issues in the frontend, making most of
the requests unusable.

Observables also proved quite challenging to handle, and unsure of the
correct syntax to use moreso in the services.

TODO: build more frontend project were you work with services,
observables,rxjs and such. Dont focus so much on styling
TODO: decipher how to properly format request and response messages from
the backend to make application work as intended.