fix(controllers): get rid of type support error
My rows column had not being strongly typed. Gave them a temporary type
of any to get rid of runtime error that appeared on all endpoints. This
makes the console cleaner when running.

Also installed jest and its types to prevent the error on the basic
test. Unsure how it had run earlier, but since i had installed the
unistaleld it, it might have been stored in cache or sth.

TODO: dicover the real type of rows from querying the db and use that
instead of any